+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-ItemCardTable
+## [0.0.14] - 2020-05-14
+### Chaged
+   - `itemCardTable` receiver title, id  props.
+   - `IconButton` new Icon Delete.
+
 ## [0.0.13] - 2020-05-10
 ### Added
    - added prop `itemCardTable`.